orderhistorystatus qoshdiim

+++ src/main/java/com/example/backend/dto/OrderItemResponse.java
@@ -3,6 +3,8 @@ package com.example.backend.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class OrderItemResponse {
@@ -10,6 +12,7 @@ public class OrderItemResponse {
     private Long productId;
     private String name;
     private String imageUrl;
-    private double price;
+    // 💰 BIGDECIMAL (ENG MUHIM FIX)
+    private BigDecimal price;
     private int quantity;
 }
