+++ src/main/java/com/example/backend/dto/ProductCreateRequest.java
@@ -1,6 +1,8 @@
 package com.example.backend.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -9,8 +11,8 @@ public class ProductCreateRequest {
     private String name;
     private String description;
     private String brand;
-    private double price;
-    private double discountPrice;
+    private BigDecimal price;
+    private BigDecimal discountPrice;
     private String category;
     private int stock;
