+++ src/main/java/com/example/backend/dto/ProductResponse.java
@@ -4,6 +4,7 @@ import com.example.backend.entity.Category;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,8 +14,11 @@ public class ProductResponse {
     private Long id;
     private String name;
     private String brand;
-    private double price;
-    private double discountPrice;
+    // 💰 BIGDECIMAL (ENG MUHIM FIX)
+    private BigDecimal price;
+    private BigDecimal discountPrice;
+
+
     private Category category;
     private double ratingAvg;
     private int reviewCount;
